requirements.txt
flask
redis
elasticsearch
psycopg2-binary
prometheus_client 
flask-limiter
kafka-python